line_bot/requirements.txt
fastapi
uvicorn
line-bot-sdk
jinja2
paho-mqtt